Revert "Refactor: memory-efficient geo file decoder (#934)"
This reverts commit 7e876709
