+++ app/router/condition_test.go
@@ -6,11 +6,13 @@ import (
 	"os"
 	"path/filepath"
 	"strconv"
+	"strings"
 	"testing"
+
+	"google.golang.org/protobuf/proto"
 
 	"github.com/v2fly/v2ray-core/v4/app/router"
 	"github.com/v2fly/v2ray-core/v4/common"
-	"github.com/v2fly/v2ray-core/v4/common/geodata"
 	"github.com/v2fly/v2ray-core/v4/common/net"
 	"github.com/v2fly/v2ray-core/v4/common/platform"
 	"github.com/v2fly/v2ray-core/v4/common/platform/filesystem"
@@ -350,8 +352,26 @@ func TestRoutingRule(t *testing.T) {
 	}
 }
 
+func loadGeoSite(country string) ([]*router.Domain, error) {
+	geositeBytes, err := filesystem.ReadAsset("geosite.dat")
+	if err != nil {
+		return nil, err
+	}
+	var geositeList router.GeoSiteList
+	if err := proto.Unmarshal(geositeBytes, &geositeList); err != nil {
+		return nil, err
+	}
+
+	for _, site := range geositeList.Entry {
+		if strings.EqualFold(site.CountryCode, country) {
+			return site.Domain, nil
+		}
+	}
+
+	return nil, errors.New("country not found: " + country)
+}
 func TestChinaSites(t *testing.T) {
-	domains, err := geodata.LoadSite("geosite.dat", "CN")
+	domains, err := loadGeoSite("CN")
 	common.Must(err)
 
 	matcher, err := router.NewDomainMatcher(domains)
@@ -398,7 +418,7 @@ func TestChinaSites(t *testing.T) {
 }
 
 func BenchmarkMphDomainMatcher(b *testing.B) {
-	domains, err := geodata.LoadSite("geosite.dat", "CN")
+	domains, err := loadGeoSite("CN")
 	common.Must(err)
 
 	matcher, err := router.NewMphMatcherGroup(domains)
@@ -440,7 +460,7 @@ func BenchmarkMphDomainMatcher(b *testing.B) {
 }
 
 func BenchmarkDomainMatcher(b *testing.B) {
-	domains, err := geodata.LoadSite("geosite.dat", "CN")
+	domains, err := loadGeoSite("CN")
 	common.Must(err)
 
 	matcher, err := router.NewDomainMatcher(domains)
@@ -485,7 +505,7 @@ func BenchmarkMultiGeoIPMatcher(b *testing.B) {
 	var geoips []*router.GeoIP
 
 	{
-		ips, err := geodata.LoadIP("geoip.dat", "CN")
+		ips, err := loadGeoIP("CN")
 		common.Must(err)
 		geoips = append(geoips, &router.GeoIP{
 			CountryCode: "CN",
@@ -494,7 +514,7 @@ func BenchmarkMultiGeoIPMatcher(b *testing.B) {
 	}
 
 	{
-		ips, err := geodata.LoadIP("geoip.dat", "JP")
+		ips, err := loadGeoIP("JP")
 		common.Must(err)
 		geoips = append(geoips, &router.GeoIP{
 			CountryCode: "JP",
@@ -503,7 +523,7 @@ func BenchmarkMultiGeoIPMatcher(b *testing.B) {
 	}
 
 	{
-		ips, err := geodata.LoadIP("geoip.dat", "CA")
+		ips, err := loadGeoIP("CA")
 		common.Must(err)
 		geoips = append(geoips, &router.GeoIP{
 			CountryCode: "CA",
@@ -512,7 +532,7 @@ func BenchmarkMultiGeoIPMatcher(b *testing.B) {
 	}
 
 	{
-		ips, err := geodata.LoadIP("geoip.dat", "US")
+		ips, err := loadGeoIP("US")
 		common.Must(err)
 		geoips = append(geoips, &router.GeoIP{
 			CountryCode: "US",
