+++ infra/conf/observatory.go
@@ -2,7 +2,6 @@ package conf
 
 import (
 	"github.com/golang/protobuf/proto"
-
 	"github.com/v2fly/v2ray-core/v4/app/observatory"
 )
